✨🧬 Added the `IconLoader` and `Button` as custom components, refactoring the folder structure as per `Atomic Design` guidelines
Why
- To implement the `Atomic Design` guidelines and better handle the SVG `icons`

How
- Changed the `tsconfing.vitest.json` file adding the `moduleResolution` of `node`
- First attempt to organise components according to `Atomic Design` guidelines
- Added the necessary `icons` to the `assets` folder
- Added the `IconLoader` atom to better handle the SVG `icons`
- Added a custom `Button` molecule (provisory)

Tags
 Changes to be committed:
	modified:   src/App.vue
	new file:   src/assets/icons/Dot.vue
	new file:   src/assets/icons/LeftArrow.vue
	new file:   src/assets/icons/RightArrow.vue
	deleted:    src/components/Calendar/CalendarDay/index.ts
	deleted:    src/components/Calendar/CalendarHours/index.ts
	deleted:    src/components/Calendar/CalendarView/index.ts
	renamed:    src/components/Calendar/Headers/DayHeader/DayHeader.vue -> src/components/atoms/DayHeader/DayHeader.vue
	renamed:    src/components/Calendar/Headers/DayHeader/index.ts -> src/components/atoms/DayHeader/index.ts
	renamed:    src/components/Events/EventCard/EventCard.vue -> src/components/atoms/EventCard/EventCard.vue
	renamed:    src/components/Events/EventCard/index.ts -> src/components/atoms/EventCard/index.ts
	renamed:    src/components/Calendar/Headers/EventsHeader/EventsHeader.vue -> src/components/atoms/EventsHeader/EventsHeader.vue
	renamed:    src/components/Calendar/Headers/EventsHeader/index.ts -> src/components/atoms/EventsHeader/index.ts
	renamed:    src/components/Calendar/CalendarHours/CalendarHours.vue -> src/components/atoms/Hours/Hours.vue
	new file:   src/components/atoms/Hours/index.ts
	new file:   src/components/atoms/IconLoader/IconLoader.vue
	new file:   src/components/atoms/IconLoader/index.ts
	new file:   src/components/molecules/Button/Button.vue
	new file:   src/components/molecules/Button/index.ts
	renamed:    src/components/Calendar/CalendarDay/CalendarDay.vue -> src/components/molecules/Day/Day.vue
	new file:   src/components/molecules/Day/index.ts
	renamed:    src/components/Calendar/CalendarView/CalendarView.vue -> src/components/organisms/Calendar/Calendar.vue
	new file:   src/components/organisms/Calendar/index.ts
	renamed:    src/components/Sidebar/Sidebar.vue -> src/components/organisms/Sidebar/Sidebar.vue
	renamed:    src/components/Sidebar/index.ts -> src/components/organisms/Sidebar/index.ts
	deleted:    src/views/Calendar.vue
	renamed:    src/views/Calendar.spec.ts -> src/views/pages/DefaultLayout/DefaultLayout.spec.ts
	new file:   src/views/pages/DefaultLayout/DefaultLayout.vue
	modified:   tsconfig.vitest.json
