--- src/components/Calendar/CalendarDay/index.ts
@@ -1,2 +1,0 @@
-import CalendarDay from './CalendarDay.vue';
-export default CalendarDay;
