--- src/components/Calendar/CalendarHours/index.ts
@@ -1,2 +1,0 @@
-import CalendarHours from './CalendarHours.vue';
-export default CalendarHours;
