--- src/components/Calendar/CalendarView/index.ts
@@ -1,2 +1,0 @@
-import CalendarView from './CalendarView.vue';
-export default CalendarView;
