+++ src/components/atoms/EventCard/EventCard.vue
