+++ src/components/atoms/EventCard/index.ts
