+++ src/components/atoms/EventsHeader/EventsHeader.vue
