+++ src/components/atoms/Hours/index.ts
@@ -1,0 +1,2 @@
+import Hours from './Hours.vue';
+export default Hours;
