+++ src/components/atoms/IconLoader/IconLoader.vue
@@ -1,0 +1,17 @@
+<template>
+  <component :is="dynamicComponent" />
+</template>
+
+<script setup lang="ts">
+import { defineAsyncComponent, computed } from 'vue';
+
+const props = defineProps<{
+  name: string;
+}>();
+
+const dynamicComponent = computed(() => {
+  return defineAsyncComponent(
+    () => import(`../../../assets/icons/${props.name}.vue`)
+  );
+});
+</script>
