+++ src/components/atoms/IconLoader/index.ts
@@ -1,0 +1,2 @@
+import IconLoader from './IconLoader.vue';
+export default IconLoader;
