+++ src/components/molecules/Button/Button.vue
@@ -1,0 +1,15 @@
+<template>
+  <button>
+    <slot>
+      <IconLoader v-if="icon" :name="icon" />
+    </slot>
+  </button>
+</template>
+
+<script setup lang="ts">
+import IconLoader from '@/components/atoms/IconLoader';
+
+defineProps<{
+  icon?: string;
+}>();
+</script>
