+++ src/components/molecules/Day/index.ts
@@ -1,0 +1,2 @@
+import Day from './Day.vue';
+export default Day;
