+++ src/components/organisms/Calendar/index.ts
@@ -1,0 +1,2 @@
+import Calendar from './Calendar.vue';
+export default Calendar;
