--- src/views/Calendar.vue
@@ -1,102 +1,0 @@
-<template>
-  <div
-    class="grid h-full w-full grid-cols-12 grid-rows-[repeat(12,minmax(0,1fr))] overflow-hidden"
-  >
-    <Sidebar />
-    <!-- Provisory -->
-    <div
-      class="col-start-4 col-end-12 row-span-2 flex flex-col items-start justify-center gap-4"
-    >
-      <div class="flex items-end gap-4 text-white">
-        <p class="text-3xl font-bold leading-7">{{ month }}</p>
-        <div class="flex items-center gap-2">
-          <button
-            class="rounded-full bg-white p-1 text-black shadow-md hover:bg-gray-200 active:translate-y-[1px] active:bg-amber-400 active:shadow-sm"
-            @click="moveBackwards"
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="14"
-              height="14"
-              viewBox="0 0 24 24"
-            >
-              <path
-                fill="currentColor"
-                d="M15.125 21.1L6.7 12.7q-.15-.15-.212-.325q-.063-.175-.063-.375t.063-.375q.062-.175.212-.325l8.425-8.425q.35-.35.875-.35t.9.375q.375.375.375.875t-.375.875L9.55 12l7.35 7.35q.35.35.35.862q0 .513-.375.888t-.875.375q-.5 0-.875-.375Z"
-              />
-            </svg>
-          </button>
-          <div class="text-white">
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="8"
-              height="8"
-              viewBox="0 0 24 24"
-            >
-              <path
-                fill="currentColor"
-                d="M12 22q-2.075 0-3.9-.788q-1.825-.787-3.175-2.137q-1.35-1.35-2.137-3.175Q2 14.075 2 12t.788-3.9q.787-1.825 2.137-3.175q1.35-1.35 3.175-2.138Q9.925 2 12 2t3.9.787q1.825.788 3.175 2.138q1.35 1.35 2.137 3.175Q22 9.925 22 12t-.788 3.9q-.787 1.825-2.137 3.175q-1.35 1.35-3.175 2.137Q14.075 22 12 22Zm0-2q3.35 0 5.675-2.325Q20 15.35 20 12q0-3.35-2.325-5.675Q15.35 4 12 4Q8.65 4 6.325 6.325Q4 8.65 4 12q0 3.35 2.325 5.675Q8.65 20 12 20Zm0 0q-3.35 0-5.675-2.325Q4 15.35 4 12q0-3.35 2.325-5.675Q8.65 4 12 4q3.35 0 5.675 2.325Q20 8.65 20 12q0 3.35-2.325 5.675Q15.35 20 12 20Z"
-              />
-            </svg>
-          </div>
-          <button
-            class="rounded-full bg-white p-1 text-black shadow-md hover:bg-gray-200 active:translate-y-[1px] active:bg-amber-400 active:shadow-sm"
-            @click="moveForward"
-          >
-            <svg
-              xmlns="http://www.w3.org/2000/svg"
-              width="14"
-              height="14"
-              viewBox="0 0 24 24"
-            >
-              <path
-                fill="currentColor"
-                d="M7.15 21.1q-.375-.375-.375-.888q0-.512.375-.887L14.475 12l-7.35-7.35q-.35-.35-.35-.875t.375-.9q.375-.375.888-.375q.512 0 .887.375l8.4 8.425q.15.15.213.325q.062.175.062.375t-.062.375q-.063.175-.213.325L8.9 21.125q-.35.35-.862.35q-.513 0-.888-.375Z"
-              />
-            </svg>
-          </button>
-        </div>
-      </div>
-      <div class="flex w-full items-center justify-between text-white">
-        <p class="w-2/5 text-xs">
-          Here are all your planned events. You will find information for each
-          event and can plan a new one as well
-        </p>
-        <button
-          class="rounded-lg bg-white px-6 py-2 font-bold text-black shadow-md hover:bg-gray-200 active:translate-y-[1px] active:bg-amber-400 active:shadow-sm"
-        >
-          Add Event
-        </button>
-      </div>
-    </div>
-    <!-- Provisory -->
-    <CalendarView />
-  </div>
-</template>
-
-<script lang="ts" setup>
-import CalendarView from '@/components/Calendar/CalendarView';
-import Sidebar from '@/components/Sidebar';
-import { computed } from 'vue';
-import { useCalendarStore } from '@/stores/calendarStore';
-import { getCurrentDates, setCurrentDates } from '@/utils/Dates';
-
-const store = useCalendarStore();
-store.view = getCurrentDates();
-
-// TODO: Rework this as an external snippet
-let month = computed(() => {
-  const getWeekStart = new Date(store.view[0]).toDateString();
-  let shortMonth = getWeekStart.slice(4, 7);
-  let year = getWeekStart.slice(-4);
-  return `${shortMonth}' ${year}`;
-});
-
-function moveForward() {
-  store.view = setCurrentDates(store.view, true);
-}
-
-function moveBackwards() {
-  store.view = setCurrentDates(store.view, false);
-}
-</script>
