+++ tsconfig.vitest.json
@@ -4,6 +4,7 @@
   "compilerOptions": {
     "composite": true,
     "lib": [],
-    "types": ["node", "jsdom", "vitest/globals"]
+    "types": ["node", "jsdom", "vitest/globals"],
+    "moduleResolution": "node"
   }
 }
